Sanctum Authorization added

+++ database/migrations/2024_06_02_120307_create_accounts_table.php
@@ -11,14 +11,16 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('accounts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('account_name');
-            $table->string('account_type');
-            $table->decimal('balance', 10, 2)->default(0.00);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('accounts')) {
+            Schema::create('accounts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('account_name');
+                $table->string('account_type');
+                $table->decimal('balance', 10, 2)->default(0.00);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
